pyproject.toml
[tool.poetry]
name = "buff"
version = "0.1.0"
description = ""
authors = ["Punit Arani <punitsai36@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11 <3.12"
aiocache = {extras = ["redis"], version = "^0.12.2"}
aiolimiter = "^1.1.0"
httpx = "^0.25.1"
motor = "^3.3.1"
pydantic = "^2.4.2"
python-dotenv = "^1.0.0"
streamlit = "^1.28.1"
tenacity = "^8.2.3"
tqdm = "^4.66.1"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"
pre-commit = "^3.4.0"
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
ruff = "^0.0.291"


[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
use_parentheses = true
include_trailing_comma = true
ensure_newline_before_comments = true
multi_line_output = 3
force_grid_wrap = 0
known_first_party = "buff"

[tool.ruff]
target-version = "py311"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
